requirements.txt
streamlit>=1.34
python-dotenv>=1.0.1
tiktoken>=0.7.0
pypdf>=4.2.0

# LangChain + LangGraph + OpenAI providers
langchain>=0.2.14
langchain-openai>=0.2.0
langgraph>=0.2.34

# Vector store
faiss-cpu>=1.8.0
